refactor: cleanup 0.3.x deprecations
