--- extensions/control-plane/api/management-api/contract-negotiation-api/src/main/java/org/eclipse/edc/connector/controlplane/api/management/contractnegotiation/model/ContractOfferDescription.java
@@ -1,79 +1,0 @@
-/*
- *  Copyright (c) 2020 - 2022 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - improvements
- *
- */
-
-package org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.model;
-
-import org.eclipse.edc.policy.model.Policy;
-
-import static org.eclipse.edc.spi.constants.CoreConstants.EDC_NAMESPACE;
-
-@Deprecated(since = "0.3.2")
-public class ContractOfferDescription {
-
-    public static final String CONTRACT_OFFER_DESCRIPTION_TYPE = EDC_NAMESPACE + "ContractOfferDescription";
-    public static final String OFFER_ID = EDC_NAMESPACE + "offerId";
-    public static final String ASSET_ID = EDC_NAMESPACE + "assetId";
-    public static final String POLICY = EDC_NAMESPACE + "policy";
-
-    private String offerId;
-    private String assetId;
-    private Policy policy;
-
-    private ContractOfferDescription() {
-    }
-
-    public String getOfferId() {
-        return offerId;
-    }
-
-    public String getAssetId() {
-        return assetId;
-    }
-
-    public Policy getPolicy() {
-        return policy;
-    }
-
-    public static final class Builder {
-        private final ContractOfferDescription description;
-
-        private Builder() {
-            description = new ContractOfferDescription();
-        }
-
-        public static ContractOfferDescription.Builder newInstance() {
-            return new ContractOfferDescription.Builder();
-        }
-
-        public ContractOfferDescription.Builder offerId(String offerId) {
-            description.offerId = offerId;
-            return this;
-        }
-
-        public ContractOfferDescription.Builder assetId(String assetId) {
-            description.assetId = assetId;
-            return this;
-        }
-
-        public ContractOfferDescription.Builder policy(Policy policy) {
-            description.policy = policy;
-            return this;
-        }
-
-        public ContractOfferDescription build() {
-            return description;
-        }
-    }
-}
