+++ extensions/control-plane/api/management-api/transfer-process-api/src/main/java/org/eclipse/edc/connector/controlplane/api/management/transferprocess/TransferProcessApiExtension.java
@@ -73,7 +73,7 @@ public class TransferProcessApiExtension implements ServiceExtension {
         managementApiTransformerRegistry.register(new JsonObjectToSuspendTransferTransformer());
         managementApiTransformerRegistry.register(new JsonObjectToTransferRequestTransformer());
 
-        validatorRegistry.register(TRANSFER_REQUEST_TYPE, TransferRequestValidator.instance(context.getMonitor()));
+        validatorRegistry.register(TRANSFER_REQUEST_TYPE, TransferRequestValidator.instance());
         validatorRegistry.register(TERMINATE_TRANSFER_TYPE, TerminateTransferValidator.instance());
 
         var newController = new TransferProcessApiController(context.getMonitor(), service, managementApiTransformerRegistry, validatorRegistry);
