+++ extensions/control-plane/api/management-api/transfer-process-api/src/test/java/org/eclipse/edc/connector/controlplane/api/management/transferprocess/validation/TransferRequestValidatorTest.java
@@ -17,7 +17,6 @@ package org.eclipse.edc.connector.controlplane.api.management.transferprocess.va
 import jakarta.json.Json;
 import jakarta.json.JsonArrayBuilder;
 import jakarta.json.JsonObject;
-import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.validator.spi.ValidationFailure;
 import org.eclipse.edc.validator.spi.Validator;
 import org.eclipse.edc.validator.spi.Violation;
@@ -28,7 +27,6 @@ import static jakarta.json.Json.createObjectBuilder;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.InstanceOfAssertFactories.list;
 import static org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferRequest.TRANSFER_REQUEST_ASSET_ID;
-import static org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferRequest.TRANSFER_REQUEST_CONNECTOR_ADDRESS;
 import static org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferRequest.TRANSFER_REQUEST_CONTRACT_ID;
 import static org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferRequest.TRANSFER_REQUEST_COUNTER_PARTY_ADDRESS;
 import static org.eclipse.edc.connector.controlplane.transfer.spi.types.TransferRequest.TRANSFER_REQUEST_DATA_DESTINATION;
@@ -38,14 +36,10 @@ import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.ID;
 import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.VALUE;
 import static org.eclipse.edc.junit.assertions.AbstractResultAssert.assertThat;
 import static org.eclipse.edc.spi.types.domain.DataAddress.EDC_DATA_ADDRESS_TYPE_PROPERTY;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
 
 class TransferRequestValidatorTest {
 
-    private final Monitor monitor = mock();
-    private final Validator<JsonObject> validator = TransferRequestValidator.instance(monitor);
+    private final Validator<JsonObject> validator = TransferRequestValidator.instance();
 
     @Test
     void shouldSucceed_whenObjectIsValid() {
@@ -78,26 +72,6 @@ class TransferRequestValidatorTest {
         var result = validator.validate(input);
 
         assertThat(result).isSucceeded();
-    }
-
-    @Test
-    @Deprecated(since = "0.3.2")
-    void shouldSucceed_whenDeprecatedConnectorAddressIsUsed() {
-        var input = Json.createObjectBuilder()
-                .add(TRANSFER_REQUEST_CONNECTOR_ADDRESS, value("http://connector-address"))
-                .add(TRANSFER_REQUEST_CONTRACT_ID, value("contract-id"))
-                .add(TRANSFER_REQUEST_PROTOCOL, value("protocol"))
-                .add(TRANSFER_REQUEST_ASSET_ID, value("assetId"))
-                .add(TRANSFER_REQUEST_TRANSFER_TYPE, value("transferType"))
-                .add(TRANSFER_REQUEST_DATA_DESTINATION, createArrayBuilder().add(createObjectBuilder()
-                        .add(EDC_DATA_ADDRESS_TYPE_PROPERTY, value("type"))
-                ))
-                .build();
-
-        var result = validator.validate(input);
-
-        assertThat(result).isSucceeded();
-        verify(monitor).warning(anyString());
     }
 
     @Test
